Added signup page

+++ backend/src/routes.js
@@ -3,6 +3,8 @@ const dbClient = require('./dbclient')
 const express = require('express')
 const jwt = require('jsonwebtoken')
 const jwtConfig = require('./jwtConfig')
+const encryption = require('./encryption')
+const fetch = require('node-fetch')
 const router = express.Router()
 const { 
   buildGetTasksQuery,
@@ -15,21 +17,57 @@ const [masterUsername, masterPassword] = [
   'bfyXtzvsajrTQukpdKSSHxKFxA9CJxLyh6rwJyQH'
 ]
 
-router.post('/login', (req, res) => {
-  const postData = req.body;
-  const username = postData.username
-  const password = postData.password
+const captchaSecret = '6LfWLL8ZAAAAAKMgAMlPf3iv2V1FD9mTr4QzRw2m'
+
+router.post('/login', async (req, res) => {
+  const {username, password } = req.body
+
+  const user = await dbClient.getElems({table: 'users', query: {username}})
   
-  if (username !== masterUsername || password !== masterPassword) {
-    res.status(401).send('Wrong username or password')
+  if (!user.length) {
+    res.status(401).send('Username does not exist')
+    return
+  }
+  
+  const {userId, encryptedPass} = user[0]
+
+
+  if (encryption.decrypt(encryptedPass) !== password) {
+    res.status(401).send('Wrong username of password')
   } else {
-    const token = jwt.sign({username}, jwtConfig.secret, { expiresIn: jwtConfig.sessionTokenLife})
+    const token = jwt.sign({userId}, jwtConfig.secret, { expiresIn: jwtConfig.sessionTokenLife})
     const response = {
         "status": "Logged in",
         "token": token,
     }
     res.status(200).json(response)
   }
+})
+
+router.post('/signup', async (req, res) => {
+  const { name, username, password, captchaToken } = req.body
+  const headers = {
+    "Content-Type": "application/x-www-form-urlencoded"
+  }
+  const body = `secret=${captchaSecret}&response=${captchaToken}`
+  const resp = await fetch(
+    'https://www.google.com/recaptcha/api/siteverify',
+    {method: 'post', headers, body }
+  ) 
+  const json = await resp.json()
+  if (json.success && name && username && password) {
+    const encryptedPass = encryption.encrypt(password)
+    const userId = await dbClient.writeElem({table: 'users', task: {name, username, encryptedPass}})
+    res.status(200)
+    res.json({'userId': userId})
+    res.end()
+  } else {
+    res.status(400)
+    res.end()
+  }
+
+
+  
 })
 
 router.use(require('./tokenChecker'))
@@ -40,9 +78,22 @@ router.get('/status', (req, res) => {
   res.end()
 })
 
+
+router.post('/task', async (req, res) => {
+  const task = req.body
+  task.createdAt = new Date().toJSON()
+  task.updatedAt = new Date().toJSON()
+  task.doneAt = null
+  const taskId = await dbClient.writeElem({table: 'tasks', elem: task, userId: req.decoded})
+  res.status(200)
+  res.json({'taskId': taskId})
+  res.end()
+})
+
+
 router.get('/tasks', async (req, res) => {
   const query = buildGetTasksQuery(req.query)
-  const tasks = await dbClient.getElems({table: 'tasks', query})
+  const tasks = await dbClient.getElems({table: 'tasks', query, userId: req.decoded})
   res.status(200)
   res.json(tasks)
   res.end()
@@ -53,7 +104,7 @@ router.post('/task', async (req, res) => {
   task.createdAt = new Date().toJSON()
   task.updatedAt = new Date().toJSON()
   task.doneAt = null
-  const taskId = await dbClient.writeElem({table: 'tasks', task})
+  const taskId = await dbClient.writeElem({table: 'tasks', elem: task, userId: req.decoded})
   res.status(200)
   res.json({'taskId': taskId})
   res.end()
@@ -63,7 +114,7 @@ router.put('/task/:taskId', async (req, res) => {
   const taskId = req.params.taskId
   const task = req.body
   task.updatedAt = new Date().toJSON()
-  await dbClient.updateElem({table: 'tasks', task, taskId})
+  await dbClient.updateElem({table: 'tasks', elem: task, elemId: taskId, userId: req.decoded})
   res.status(200)
   res.json({'taskId': taskId})
   res.end()
@@ -71,7 +122,7 @@ router.put('/task/:taskId', async (req, res) => {
 
 router.delete('/task/:taskId', async (req, res) => {
   const taskId = req.params.taskId
-  await dbClient.deleteElem({table: 'tasks', taskId})
+  await dbClient.deleteElem({table: 'tasks', elemId: taskId, userId: req.decoded})
   res.status(200)
   res.json({'taskId': taskId})
   res.end()
@@ -79,45 +130,45 @@ router.delete('/task/:taskId', async (req, res) => {
 
 router.get('/goals', async (req, res) => {
   const query = {}
-  const goals = await dbClient.getElems({table: 'goals', query})
+  const goals = await dbClient.getElems({table: 'goals', query, userId: req.decoded})
   res.status(200)
   res.json(goals)
   res.end()
 })
 
 router.post('/goal', async (req, res) => {
-  const task = req.body
-  task.createdAt = new Date().toJSON()
-  task.updatedAt = new Date().toJSON()
-  task.colorCode = getRandomColor()
-  task.doneAt = null
-  const goalId = await dbClient.writeElem({table: 'goals', task})
+  const goal = req.body
+  goal.createdAt = new Date().toJSON()
+  goal.updatedAt = new Date().toJSON()
+  goal.colorCode = getRandomColor()
+  goal.doneAt = null
+  const goalId = await dbClient.writeElem({table: 'goals', elem: goal, userId: req.decoded})
   res.status(200)
   res.json({'goalId': goalId})
   res.end()
 })
 
-router.put('/goal/:taskId', async (req, res) => {
-  const taskId = req.params.taskId
-  const task = req.body
-  task.updatedAt = new Date().toJSON()
-  await dbClient.updateElem({table: 'goals', task, taskId})
+router.put('/goal/:goalId', async (req, res) => {
+  const goalId = req.params.goalId
+  const goal = req.body
+  goal.updatedAt = new Date().toJSON()
+  await dbClient.updateElem({table: 'goals', elem: goal, elemId: goalId, userId: req.decoded})
   res.status(200)
-  res.json({'goalId': taskId})
+  res.json({'goalId': goalId})
   res.end()
 })
 
-router.delete('/goal/:taskId', async (req, res) => {
-  const taskId = req.params.taskId
-  await dbClient.deleteElem({table: 'goals', taskId})
+router.delete('/goal/:goalId', async (req, res) => {
+  const goalId = req.params.goalId
+  await dbClient.deleteElem({table: 'goals', elemId: goalId, userId: req.decoded})
   res.status(200)
-  res.json({'goalId': taskId})
+  res.json({'goalId': goalId})
   res.end()
 })
 
 router.get('/projects', async (req, res) => {
   const query = buildGetProjectsQuery(req.query)
-  const projects = await dbClient.getElems({table: 'projects', query})
+  const projects = await dbClient.getElems({table: 'projects', query, userId: req.decoded})
   res.status(200)
   res.json(projects)
   res.end()
@@ -126,39 +177,39 @@ router.get('/projects', async (req, res) => {
 router.get('/project/:projectId/tasks', async (req, res) => {
   const projectId = req.params.projectId
   const query = {'projectId': projectId}
-  const tasks = await dbClient.getElems({table: 'tasks', query})
+  const tasks = await dbClient.getElems({table: 'tasks', query, userId: req.decoded})
   res.status(200)
   res.json(tasks)
   res.end()
 })
 
 router.post('/project', async (req, res) => {
-  const task = req.body
-  task.createdAt = new Date().toJSON()
-  task.updatedAt = new Date().toJSON()
-  task.colorCode = getRandomColor()
-  task.doneAt = null
-  const projectId = await dbClient.writeElem({table: 'projects', task})
+  const project = req.body
+  project.createdAt = new Date().toJSON()
+  project.updatedAt = new Date().toJSON()
+  project.colorCode = getRandomColor()
+  project.doneAt = null
+  const projectId = await dbClient.writeElem({table: 'projects', elem: project, userId: req.decoded})
   res.status(200)
   res.json({'projectId': projectId})
   res.end()
 })
 
-router.put('/project/:taskId', async (req, res) => {
-  const taskId = req.params.taskId
-  const task = req.body
-  task.updatedAt = new Date().toJSON()
-  await dbClient.updateElem({table: 'projects', task, taskId})
+router.put('/project/:projectId', async (req, res) => {
+  const projectId = req.params.projectId
+  const project = req.body
+  project.updatedAt = new Date().toJSON()
+  await dbClient.updateElem({table: 'projects', elem: project, elemId: projectId, userId: req.decoded})
   res.status(200)
-  res.json({'projectId': taskId})
+  res.json({'projectId': projectId})
   res.end()
 })
 
-router.delete('/project/:taskId', async (req, res) => {
-  const taskId = req.params.taskId
-  await dbClient.deleteElem({table: 'projects', taskId})
+router.delete('/project/:projectId', async (req, res) => {
+  const projectId = req.params.projectId
+  await dbClient.deleteElem({table: 'projects', elemId: projectId, userId: req.decoded})
   res.status(200)
-  res.json({'projectId': taskId})
+  res.json({'projectId': projectId})
   res.end()
 })
